working on #280 : making the code 64bit ready

+++ elenasrc2/elenasm/win32/dllmain.cpp
@@ -51,7 +51,7 @@ EXTERN_DLL_EXPORT int GetStatus(char* buffer, int maxLength)
    if (session) {
       _ELENA_::ident_t error = session->getLastError();
       size_t length = _ELENA_::getlength(error);
-      if (length > maxLength)
+      if ((int)length > maxLength)
          length = maxLength;
 
       _ELENA_::Convertor::copy(buffer, error, length, length);
